+++ assets/SanitizerConfigW3C.js
@@ -1,5 +1,5 @@
 /**
- * @copyright 2023 Chris Zuber <admin@kernvalley.us>
+ * @copyright 2023-2024 Chris Zuber <admin@kernvalley.us>
  * @see https://wicg.github.io/sanitizer-api/#default-configuration-dictionary
  */
 export const SanitizerConfigW3C = JSON.parse(`{
@@ -125,636 +125,217 @@ export const SanitizerConfigW3C = JSON.parse(`{
     "wbr"
   ],
   "allowAttributes": {
-    "abbr": [
-      "*"
-    ],
-    "accept": [
-      "*"
-    ],
-    "accept-charset": [
-      "*"
-    ],
-    "accesskey": [
-      "*"
-    ],
-    "action": [
-      "*"
-    ],
-    "align": [
-      "*"
-    ],
-    "alink": [
-      "*"
-    ],
-    "allow": [
-      "*"
-    ],
-    "allowfullscreen": [
-      "*"
-    ],
-    "alt": [
-      "*"
-    ],
-    "anchor": [
-      "*"
-    ],
-    "archive": [
-      "*"
-    ],
-    "as": [
-      "*"
-    ],
-    "async": [
-      "*"
-    ],
-    "autocapitalize": [
-      "*"
-    ],
-    "autocomplete": [
-      "*"
-    ],
-    "autocorrect": [
-      "*"
-    ],
-    "autofocus": [
-      "*"
-    ],
-    "autopictureinpicture": [
-      "*"
-    ],
-    "autoplay": [
-      "*"
-    ],
-    "axis": [
-      "*"
-    ],
-    "background": [
-      "*"
-    ],
-    "behavior": [
-      "*"
-    ],
-    "bgcolor": [
-      "*"
-    ],
-    "border": [
-      "*"
-    ],
-    "bordercolor": [
-      "*"
-    ],
-    "capture": [
-      "*"
-    ],
-    "cellpadding": [
-      "*"
-    ],
-    "cellspacing": [
-      "*"
-    ],
-    "challenge": [
-      "*"
-    ],
-    "char": [
-      "*"
-    ],
-    "charoff": [
-      "*"
-    ],
-    "charset": [
-      "*"
-    ],
-    "checked": [
-      "*"
-    ],
-    "cite": [
-      "*"
-    ],
-    "class": [
-      "*"
-    ],
-    "classid": [
-      "*"
-    ],
-    "clear": [
-      "*"
-    ],
-    "code": [
-      "*"
-    ],
-    "codebase": [
-      "*"
-    ],
-    "codetype": [
-      "*"
-    ],
-    "color": [
-      "*"
-    ],
-    "cols": [
-      "*"
-    ],
-    "colspan": [
-      "*"
-    ],
-    "compact": [
-      "*"
-    ],
-    "content": [
-      "*"
-    ],
-    "contenteditable": [
-      "*"
-    ],
-    "controls": [
-      "*"
-    ],
-    "controlslist": [
-      "*"
-    ],
-    "conversiondestination": [
-      "*"
-    ],
-    "coords": [
-      "*"
-    ],
-    "crossorigin": [
-      "*"
-    ],
-    "csp": [
-      "*"
-    ],
-    "data": [
-      "*"
-    ],
-    "datetime": [
-      "*"
-    ],
-    "declare": [
-      "*"
-    ],
-    "decoding": [
-      "*"
-    ],
-    "default": [
-      "*"
-    ],
-    "defer": [
-      "*"
-    ],
-    "dir": [
-      "*"
-    ],
-    "direction": [
-      "*"
-    ],
-    "dirname": [
-      "*"
-    ],
-    "disabled": [
-      "*"
-    ],
-    "disablepictureinpicture": [
-      "*"
-    ],
-    "disableremoteplayback": [
-      "*"
-    ],
-    "disallowdocumentaccess": [
-      "*"
-    ],
-    "download": [
-      "*"
-    ],
-    "draggable": [
-      "*"
-    ],
-    "elementtiming": [
-      "*"
-    ],
-    "enctype": [
-      "*"
-    ],
-    "end": [
-      "*"
-    ],
-    "enterkeyhint": [
-      "*"
-    ],
-    "event": [
-      "*"
-    ],
-    "exportparts": [
-      "*"
-    ],
-    "face": [
-      "*"
-    ],
-    "for": [
-      "*"
-    ],
-    "form": [
-      "*"
-    ],
-    "formaction": [
-      "*"
-    ],
-    "formenctype": [
-      "*"
-    ],
-    "formmethod": [
-      "*"
-    ],
-    "formnovalidate": [
-      "*"
-    ],
-    "formtarget": [
-      "*"
-    ],
-    "frame": [
-      "*"
-    ],
-    "frameborder": [
-      "*"
-    ],
-    "headers": [
-      "*"
-    ],
-    "height": [
-      "*"
-    ],
-    "hidden": [
-      "*"
-    ],
-    "high": [
-      "*"
-    ],
-    "href": [
-      "*"
-    ],
-    "hreflang": [
-      "*"
-    ],
-    "hreftranslate": [
-      "*"
-    ],
-    "hspace": [
-      "*"
-    ],
-    "http-equiv": [
-      "*"
-    ],
-    "id": [
-      "*"
-    ],
-    "imagesizes": [
-      "*"
-    ],
-    "imagesrcset": [
-      "*"
-    ],
-    "importance": [
-      "*"
-    ],
-    "impressiondata": [
-      "*"
-    ],
-    "impressionexpiry": [
-      "*"
-    ],
-    "incremental": [
-      "*"
-    ],
-    "inert": [
-      "*"
-    ],
-    "inputmode": [
-      "*"
-    ],
-    "integrity": [
-      "*"
-    ],
-    "invisible": [
-      "*"
-    ],
-    "is": [
-      "*"
-    ],
-    "ismap": [
-      "*"
-    ],
-    "keytype": [
-      "*"
-    ],
-    "kind": [
-      "*"
-    ],
-    "label": [
-      "*"
-    ],
-    "lang": [
-      "*"
-    ],
-    "language": [
-      "*"
-    ],
-    "latencyhint": [
-      "*"
-    ],
-    "leftmargin": [
-      "*"
-    ],
-    "link": [
-      "*"
-    ],
-    "list": [
-      "*"
-    ],
-    "loading": [
-      "*"
-    ],
-    "longdesc": [
-      "*"
-    ],
-    "loop": [
-      "*"
-    ],
-    "low": [
-      "*"
-    ],
-    "lowsrc": [
-      "*"
-    ],
-    "manifest": [
-      "*"
-    ],
-    "marginheight": [
-      "*"
-    ],
-    "marginwidth": [
-      "*"
-    ],
-    "max": [
-      "*"
-    ],
-    "maxlength": [
-      "*"
-    ],
-    "mayscript": [
-      "*"
-    ],
-    "media": [
-      "*"
-    ],
-    "method": [
-      "*"
-    ],
-    "min": [
-      "*"
-    ],
-    "minlength": [
-      "*"
-    ],
-    "multiple": [
-      "*"
-    ],
-    "muted": [
-      "*"
-    ],
-    "name": [
-      "*"
-    ],
-    "nohref": [
-      "*"
-    ],
-    "nomodule": [
-      "*"
-    ],
-    "nonce": [
-      "*"
-    ],
-    "noresize": [
-      "*"
-    ],
-    "noshade": [
-      "*"
-    ],
-    "novalidate": [
-      "*"
-    ],
-    "nowrap": [
-      "*"
-    ],
-    "object": [
-      "*"
-    ],
-    "open": [
-      "*"
-    ],
-    "optimum": [
-      "*"
-    ],
-    "part": [
-      "*"
-    ],
-    "pattern": [
-      "*"
-    ],
-    "ping": [
-      "*"
-    ],
-    "placeholder": [
-      "*"
-    ],
-    "playsinline": [
-      "*"
-    ],
-    "policy": [
-      "*"
-    ],
-    "poster": [
-      "*"
-    ],
-    "preload": [
-      "*"
-    ],
-    "pseudo": [
-      "*"
-    ],
-    "readonly": [
-      "*"
-    ],
-    "referrerpolicy": [
-      "*"
-    ],
-    "rel": [
-      "*"
-    ],
-    "reportingorigin": [
-      "*"
-    ],
-    "required": [
-      "*"
-    ],
-    "resources": [
-      "*"
-    ],
-    "rev": [
-      "*"
-    ],
-    "reversed": [
-      "*"
-    ],
-    "role": [
-      "*"
-    ],
-    "rows": [
-      "*"
-    ],
-    "rowspan": [
-      "*"
-    ],
-    "rules": [
-      "*"
-    ],
-    "sandbox": [
-      "*"
-    ],
-    "scheme": [
-      "*"
-    ],
-    "scope": [
-      "*"
-    ],
-    "scopes": [
-      "*"
-    ],
-    "scrollamount": [
-      "*"
-    ],
-    "scrolldelay": [
-      "*"
-    ],
-    "scrolling": [
-      "*"
-    ],
-    "select": [
-      "*"
-    ],
-    "selected": [
-      "*"
-    ],
-    "shadowroot": [
-      "*"
-    ],
-    "shadowrootdelegatesfocus": [
-      "*"
-    ],
-    "shape": [
-      "*"
-    ],
-    "size": [
-      "*"
-    ],
-    "sizes": [
-      "*"
-    ],
-    "slot": [
-      "*"
-    ],
-    "span": [
-      "*"
-    ],
-    "spellcheck": [
-      "*"
-    ],
-    "src": [
-      "*"
-    ],
-    "srcdoc": [
-      "*"
-    ],
-    "srclang": [
-      "*"
-    ],
-    "srcset": [
-      "*"
-    ],
-    "standby": [
-      "*"
-    ],
-    "start": [
-      "*"
-    ],
-    "step": [
-      "*"
-    ],
-    "style": [
-      "*"
-    ],
-    "summary": [
-      "*"
-    ],
-    "tabindex": [
-      "*"
-    ],
-    "target": [
-      "*"
-    ],
-    "text": [
-      "*"
-    ],
-    "title": [
-      "*"
-    ],
-    "topmargin": [
-      "*"
-    ],
-    "translate": [
-      "*"
-    ],
-    "truespeed": [
-      "*"
-    ],
-    "trusttoken": [
-      "*"
-    ],
-    "type": [
-      "*"
-    ],
-    "usemap": [
-      "*"
-    ],
-    "valign": [
-      "*"
-    ],
-    "value": [
-      "*"
-    ],
-    "valuetype": [
-      "*"
-    ],
-    "version": [
-      "*"
-    ],
-    "virtualkeyboardpolicy": [
-      "*"
-    ],
-    "vlink": [
-      "*"
-    ],
-    "vspace": [
-      "*"
-    ],
-    "webkitdirectory": [
-      "*"
-    ],
-    "width": [
-      "*"
-    ],
-    "wrap": [
-      "*"
-    ]
+    "abbr": ["*"],
+    "accept": ["*"],
+    "accept-charset": ["*"],
+    "accesskey": ["*"],
+    "action": ["*"],
+    "align": ["*"],
+    "alink": ["*"],
+    "allow": ["*"],
+    "allowfullscreen": ["*"],
+    "alt": ["*"],
+    "anchor": ["*"],
+    "archive": ["*"],
+    "as": ["*"],
+    "async": ["*"],
+    "autocapitalize": ["*"],
+    "autocomplete": ["*"],
+    "autocorrect": ["*"],
+    "autofocus": ["*"],
+    "autopictureinpicture": ["*"],
+    "autoplay": ["*"],
+    "axis": ["*"],
+    "background": ["*"],
+    "behavior": ["*"],
+    "bgcolor": ["*"],
+    "border": ["*"],
+    "bordercolor": ["*"],
+    "capture": ["*"],
+    "cellpadding": ["*"],
+    "cellspacing": ["*"],
+    "challenge": ["*"],
+    "char": ["*"],
+    "charoff": ["*"],
+    "charset": ["*"],
+    "checked": ["*"],
+    "cite": ["*"],
+    "class": ["*"],
+    "classid": ["*"],
+    "clear": ["*"],
+    "code": ["*"],
+    "codebase": ["*"],
+    "codetype": ["*"],
+    "color": ["*"],
+    "cols": ["*"],
+    "colspan": ["*"],
+    "compact": ["*"],
+    "content": ["*"],
+    "contenteditable": ["*"],
+    "controls": ["*"],
+    "controlslist": ["*"],
+    "conversiondestination": ["*"],
+    "coords": ["*"],
+    "crossorigin": ["*"],
+    "csp": ["*"],
+    "data": ["*"],
+    "datetime": ["*"],
+    "declare": ["*"],
+    "decoding": ["*"],
+    "default": ["*"],
+    "defer": ["*"],
+    "dir": ["*"],
+    "direction": ["*"],
+    "dirname": ["*"],
+    "disabled": ["*"],
+    "disablepictureinpicture": ["*"],
+    "disableremoteplayback": ["*"],
+    "disallowdocumentaccess": ["*"],
+    "download": ["*"],
+    "draggable": ["*"],
+    "elementtiming": ["*"],
+    "enctype": ["*"],
+    "end": ["*"],
+    "enterkeyhint": ["*"],
+    "event": ["*"],
+    "exportparts": ["*"],
+    "face": ["*"],
+    "for": ["*"],
+    "form": ["*"],
+    "formaction": ["*"],
+    "formenctype": ["*"],
+    "formmethod": ["*"],
+    "formnovalidate": ["*"],
+    "formtarget": ["*"],
+    "frame": ["*"],
+    "frameborder": ["*"],
+    "headers": ["*"],
+    "height": ["*"],
+    "hidden": ["*"],
+    "high": ["*"],
+    "href": ["*"],
+    "hreflang": ["*"],
+    "hreftranslate": ["*"],
+    "hspace": ["*"],
+    "http-equiv": ["*"],
+    "id": ["*"],
+    "imagesizes": ["*"],
+    "imagesrcset": ["*"],
+    "importance": ["*"],
+    "impressiondata": ["*"],
+    "impressionexpiry": ["*"],
+    "incremental": ["*"],
+    "inert": ["*"],
+    "inputmode": ["*"],
+    "integrity": ["*"],
+    "invisible": ["*"],
+    "is": ["*"],
+    "ismap": ["*"],
+    "keytype": ["*"],
+    "kind": ["*"],
+    "label": ["*"],
+    "lang": ["*"],
+    "language": ["*"],
+    "latencyhint": ["*"],
+    "leftmargin": ["*"],
+    "link": ["*"],
+    "list": ["*"],
+    "loading": ["*"],
+    "longdesc": ["*"],
+    "loop": ["*"],
+    "low": ["*"],
+    "lowsrc": ["*"],
+    "manifest": ["*"],
+    "marginheight": ["*"],
+    "marginwidth": ["*"],
+    "max": ["*"],
+    "maxlength": ["*"],
+    "mayscript": ["*"],
+    "media": ["*"],
+    "method": ["*"],
+    "min": ["*"],
+    "minlength": ["*"],
+    "multiple": ["*"],
+    "muted": ["*"],
+    "name": ["*"],
+    "nohref": ["*"],
+    "nomodule": ["*"],
+    "nonce": ["*"],
+    "noresize": ["*"],
+    "noshade": ["*"],
+    "novalidate": ["*"],
+    "nowrap": ["*"],
+    "object": ["*"],
+    "open": ["*"],
+    "optimum": ["*"],
+    "part": ["*"],
+    "pattern": ["*"],
+    "ping": ["*"],
+    "placeholder": ["*"],
+    "playsinline": ["*"],
+    "policy": ["*"],
+    "poster": ["*"],
+    "preload": ["*"],
+    "pseudo": ["*"],
+    "readonly": ["*"],
+    "referrerpolicy": ["*"],
+    "rel": ["*"],
+    "reportingorigin": ["*"],
+    "required": ["*"],
+    "resources": ["*"],
+    "rev": ["*"],
+    "reversed": ["*"],
+    "role": ["*"],
+    "rows": ["*"],
+    "rowspan": ["*"],
+    "rules": ["*"],
+    "sandbox": ["*"],
+    "scheme": ["*"],
+    "scope": ["*"],
+    "scopes": ["*"],
+    "scrollamount": ["*"],
+    "scrolldelay": ["*"],
+    "scrolling": ["*"],
+    "select": ["*"],
+    "selected": ["*"],
+    "shadowroot": ["*"],
+    "shadowrootmode": ["*"],
+    "shadowrootdelegatesfocus": ["*"],
+    "shape": ["*"],
+    "size": ["*"],
+    "sizes": ["*"],
+    "slot": ["*"],
+    "span": ["*"],
+    "spellcheck": ["*"],
+    "src": ["*"],
+    "srcdoc": ["*"],
+    "srclang": ["*"],
+    "srcset": ["*"],
+    "standby": ["*"],
+    "start": ["*"],
+    "step": ["*"],
+    "style": ["*"],
+    "summary": ["*"],
+    "tabindex": ["*"],
+    "target": ["*"],
+    "text": ["*"],
+    "title": ["*"],
+    "topmargin": ["*"],
+    "translate": ["*"],
+    "truespeed": ["*"],
+    "trusttoken": ["*"],
+    "type": ["*"],
+    "usemap": ["*"],
+    "valign": ["*"],
+    "value": ["*"],
+    "valuetype": ["*"],
+    "version": ["*"],
+    "virtualkeyboardpolicy": ["*"],
+    "vlink": ["*"],
+    "vspace": ["*"],
+    "webkitdirectory": ["*"],
+    "width": ["*"],
+    "wrap": ["*"]
   }
 }`);